+++ src/main/java/uk/gov/hmcts/reform/migration/domain/OldRespondent.java
@@ -1,10 +1,10 @@
-package uk.gov.hmcts.reform.fpl.domain;
+package uk.gov.hmcts.reform.migration.domain;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
 import lombok.Data;
-import uk.gov.hmcts.reform.fpl.domain.common.Address;
+import uk.gov.hmcts.reform.migration.domain.common.Address;
 
 @Data
 @Builder
